Created the Main Menu buttons with functionality
The Play button takes you to the next scene (a placeholder for now), the About button brings up a new panel with a Back button (no text yet), the Instructinos button brings up a placeholder text and a Back button, and the Quit button has a text box to confirm you wish to quit then you can quit or go back. I need to add text and make it pretty. I also want to add the hospital image, but I have one in the resources section.