Updated MainMenu Scene 01 and Game Instructions Scene 02
Moved the first About panel text information to a new text on the opening scene (01) Main Manu to provide initial 'about' information when the game opens.

Deleted the original first panel of scene 02, added part of that text to the opening image of scene 02 for context before any buttons get clicked.